requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2

